+++ myapp.vcxproj
@@ -121,7 +121,8 @@
   </ItemDefinitionGroup>
   <!-- END Custom section -->
   <ItemGroup>
-    <ClCompile Include="myapp.cpp" />
+    <ClCompile Include="rr_app.cpp" />
+    <ClCompile Include="src\raytracing\raytracer.cpp" />
     <ClCompile Include="src\camera.cpp" />
     <ClCompile Include="src\scene_manager.cpp" />
     <ClCompile Include="template\template.cpp">
@@ -132,6 +133,7 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="src\raytracing\raytracer.h" />
     <ClInclude Include="src\lights\ambient_light.h" />
     <ClInclude Include="src\camera.h" />
     <ClInclude Include="src\materials\checkerboard_material.h" />
@@ -144,13 +146,13 @@
     <ClInclude Include="src\materials\normal_material.h" />
     <ClInclude Include="src\lights\point_light.h" />
     <ClInclude Include="src\materials\diffuse_material.h" />
-    <ClInclude Include="src\hittable.h" />
+    <ClInclude Include="src\raytracing\hittable.h" />
     <ClInclude Include="src\lights\light.h" />
     <ClInclude Include="src\materials\material.h" />
-    <ClInclude Include="myapp.h" />
+    <ClInclude Include="rr_app.h" />
     <ClInclude Include="src\3d-objects\object3d.h" />
     <ClInclude Include="src\3d-objects\primitives\plane.h" />
-    <ClInclude Include="src\ray.h" />
+    <ClInclude Include="src\raytracing\ray.h" />
     <ClInclude Include="src\scene.h" />
     <ClInclude Include="src\3d-objects\primitives\sphere.h" />
     <ClInclude Include="template\common.h" />
